// src/app/washing-machine/models/endpoints/get-washing-machine-full.endpoint.ts
import { DamageType } from '../../enums/damage-type.enum';
import { IdentificationMode } from '../../enums/identification-mode.enum';
import { Recommendation } from '../../enums/recommendation.enum';
import { ReturnType } from '../../enums/return-type.enum';

export interface GetWashingMachineFullResponse {
  category: string;
  identificationMode: IdentificationMode;
  manufacturer: string;

  model: string;
  type: string;
  serialNumber: string;

  returnType: ReturnType;
  damageType: DamageType;

  recommendation: Recommendation;
  createdAt: Date;

  washingMachineDetail?: WashingMachineDetail;
  washingMachineImages?: WashingMachineImage[];
}

interface WashingMachineDetail {
  // PACKAGE
  applicablePackageDamage: boolean;
  packageDamaged: boolean;
  packageDirty: boolean;
  packageMaterialAvailable: boolean;

  // VISIBLE SURFACES
  applicableVisibleSurfacesDamage: boolean;

  visibleSurfacesHasScratches: boolean;
  visibleSurfacesScratchesLength: number;

  visibleSurfacesHasDents: boolean;
  visibleSurfacesDentsDepth: number;

  visibleSurfacesHasMinorDamage: boolean;
  visibleSurfacesMinorDamage: string;

  visibleSurfacesHasMajorDamage: boolean;
  visibleSurfacesMajorDamage: string;

  // HIDDEN SURFACES
  applicableHiddenSurfacesDamage: boolean;

  hiddenSurfacesHasScratches: boolean;
  hiddenSurfacesScratchesLength: number;

  hiddenSurfacesHasDents: boolean;
  hiddenSurfacesDentsDepth: number;

  hiddenSurfacesHasMinorDamage: boolean;
  hiddenSurfacesMinorDamage: string;

  hiddenSurfacesHasMajorDamage: boolean;
  hiddenSurfacesMajorDamage: string;

  // PRICING
  price: number;
  repairPrice: number;
}

interface WashingMachineImage {
  imagePrefix: string;
  image: string;
}